Fix cross-plan integration issues after merging all 6 plans
- Update test shallow vecs from u32 to u64 in retained.rs and dominators.rs
  (plan 002 and 003 tests predated plan 004's u32→u64 type change)
- Remove now-no-op `as u64` casts from graph.shallow[v] accesses in
  retained.rs and graph.rs (values are already u64 after plan 004)

+++ src/graph.rs
@@ -216,7 +216,7 @@ impl ObjectGraph {
                 |mut acc, i| {
                     let c = self.object_class[i] as usize;
                     acc.0[c] += 1;
-                    acc.1[c] += self.shallow[i] as u64;
+                    acc.1[c] += self.shallow[i];
                     acc
                 },
             )
